Add retention service auto-configuration

+++ idempotence4j-core/src/main/java/com/transferwise/idempotence4j/core/ActionRepository.java
@@ -7,5 +7,5 @@ public interface ActionRepository {
     Optional<Action> find(ActionId actionId);
     Action insertOrGet(Action action);
     Action update(Action action);
-    void deleteOlderThan(Instant timestamp, Long batchSize);
+    void deleteOlderThan(Instant timestamp, Integer batchSize);
 }
