+++ idempotence4j-postgres/src/main/java/com/transferwise/idempotence4j/postgres/JdbcPostgresActionRepository.java
@@ -50,7 +50,7 @@ public class JdbcPostgresActionRepository implements ActionRepository {
 	}
 
     @Override
-    public void deleteOlderThan(Instant timestamp, Long batchSize) {
+    public void deleteOlderThan(Instant timestamp, Integer batchSize) {
         MapSqlParameterSource queryParameters = new MapSqlParameterSource()
             .addValue("createdAt", Timestamp.from(timestamp))
             .addValue("limit", batchSize);
